Minor: moved file

+++ oving4/code/summation_openmp.cpp
@@ -12,15 +12,19 @@ int main(int argc, char **argv)
     std::vector<double> vec_results ( 11 );
     std::vector<double> vec ( (int) pow(2, 14) );
     
+    // Create the vector
     for(double i=1; i<=vec.size(); i++ )
         vec[i-1]= 1/(i*i);
     
+    // Loop over the specified n values
     for( int i=3; i<=14; i++){
-        k = (int) pow(2, i);
+        // Perform the calculation
+        k = pow(2, i);
         tot = 0.0;
         #pragma omp parallel for schedule(static) reduction(+:tot)
         for( int j=0; j<k; j++ )
             tot += vec[j];
+        
         vec_results[i-3] = tot;
     }
